+++ imagery/i.atcorr/gauss.h
@@ -10,17 +10,17 @@ const long int np	= 49;
 struct Gauss
 {
 private:
-	static float angmu[10];
-	static float angphi[13];
+	static double angmu[10];
+	static double angphi[13];
 
 public:
 	/* [a,b] = [0,2*Pi] */
-	float rp[np];			/* gaussian angles */
-	float gp[np];			/* gaussian weights */
+	double rp[np];			/* gaussian angles */
+	double gp[np];			/* gaussian weights */
 
 	// [a,b] = [-1,1]
-	float rm[2*mu+1];		/* shifted gaussian angles */
-	float gb[2*mu+1];		/* shifted gaussian weights */
+	double rm[2*mu+1];		/* shifted gaussian angles */
+	double gb[2*mu+1];		/* shifted gaussian weights */
 					/* with the ends zeroed as well as the center */
 					/* [0 ? ? ? ? 0 ? ? ? ? 0] */
 
@@ -30,7 +30,7 @@ public:
 	/*	Compute for a given n, the gaussian quadrature (the n gaussian angles and the
 	their respective weights). The gaussian quadrature is used in numerical integration involving the
 	cosine of emergent or incident direction zenith angle. */
-	static void gauss (float a, float b, float *x, float *w, long int n);
+	static void gauss (double a, double b, double *x, double *w, long int n);
 };
 
 #endif /* MY_GAUSS_H */
